Move some common functions to utils.py
