+++ src/ingest-pipeline/airflow/dags/salmon_rnaseq_10x.py
@@ -1,8 +1,6 @@
-import sys
 import os
 import json
 import shlex
-from pathlib import Path
 from pprint import pprint
 from datetime import datetime, timedelta
 
@@ -18,31 +16,15 @@ import utils
 from utils import (
     PIPELINE_BASE_DIR,
     find_pipeline_manifests,
+    get_dataset_uuid,
+    get_parent_dataset_uuid,
+    get_uuid_for_error,
     localized_assert_json_matches_schema as assert_json_matches_schema,
 )
 
 import cwltool  # used to find its path
 
 THREADS = 6  # to be used by the CWL worker
-
-
-def get_parent_dataset_uuid(**kwargs):
-    return kwargs['dag_run'].conf['parent_submission_id']
-
-
-def get_dataset_uuid(**kwargs):
-    return kwargs['ti'].xcom_pull(key='derived_dataset_uuid',
-                                  task_ids="send_create_dataset")
-
-
-def get_uuid_for_error(**kwargs):
-    """
-    Return the uuid for the derived dataset if it exists, and of the parent dataset otherwise.
-    """
-    rslt = get_dataset_uuid(**kwargs)
-    if rslt is None:
-        rslt = get_parent_dataset_uuid(**kwargs)
-    return rslt
 
 
 default_args = {
